o  fix a recursive method call caused by refactoring.

+++ plexus-components/trunk/plexus-jdo2/src/main/java/org/codehaus/plexus/jdo/PlexusJdoUtils.java
@@ -271,7 +271,7 @@ public class PlexusJdoUtils
     public static Object getObjectById( PersistenceManager pm, Class clazz, int id, String fetchGroup )
         throws PlexusStoreException, PlexusObjectNotFoundException
     {
-        return getObjectById( pm, clazz, id, fetchGroup );
+        return getObjectById( pm, clazz, (long)id, fetchGroup );
     }
 
     /**
